pyproject.toml
[project]
name = "todoapp-sql-database"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.1",
    "fastapi>=0.115.8",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]
